Move EllaSparseItem to an imported file
